Add Board class
