+++ lib/board.rb
@@ -1,0 +1,11 @@
+class Board
+  def board(arr)
+    puts '+---+---+---+'
+    puts "| #{arr[0]} | #{arr[1]} | #{arr[2]} |"
+    puts '+---+---+---+'
+    puts "| #{arr[3]} | #{arr[4]} | #{arr[5]} |"
+    puts '+---+---+---+'
+    puts "| #{arr[6]} | #{arr[7]} | #{arr[8]} |"
+    puts '+---+---+---+'
+  end
+end
